requirements.txt
langchain
langchain_google_genai
python-dotenv
streamlit
gtts
pydub
